ajustes no servidor ws

+++ main.go
@@ -184,7 +184,7 @@ func (p *Player) canMoveTo(toX, toY, toDirection int) bool {
 	lastMovementMS := p.LastMovementTime.UnixNano() / int64(time.Millisecond)
 	ms := currentMS - lastMovementMS
 
-	if ms <= int(p.MovementDelay) {
+	if ms <= int64(p.MovementDelay) {
 		return false
 	}
